perf(webui): 离屏栅格宽度加上限, 把设备像素比从帧缓冲上解耦
症状: 游戏本体 700fps, 而平板里页面自身的动画只有约 15fps。

成因不在 MC 的渲染线程, 在 CEF 自己的出帧。面板的离屏表面原本严格等于它在屏幕上占的帧缓冲像素
(设备像素比恒 1:1), 4K 上是 2688x1439 = 387 万像素; 而 CEF 的离屏渲染没有共享贴图通道, 这一整张
要在 CPU 上栅格与合成完再交出来。CEF 的离屏出帧上限本就是 30fps, 单帧超预算就退化成隔一拍出一帧,
正好落在观察到的 15fps 上。

新增 webui.maxRenderWidth (默认 1600), 只在面板宽度超过它时介入, 且介入的是设备像素比而不是
CSS 视口: 缩表面的同时按同一比例缩 CEF zoom, 页面拿到的 CSS 视口逐像素不变, 响应式断点与排版
一个都不动, 变的只有贴图采样密度 (按 GL_LINEAR 放大回面板尺寸, 文字糊一档)。
4K 默认覆盖比例下像素量降到 35%; 1080p 算出来是 1344, 低于上限, 一点画质都不损失。

setZoomPercent 随之收 double —— 传进去的已经不是配置里那个整数, 而是配置值乘降采样比例。

+++ src/main/java/com/miningdim/client/webui/WebBrowser.java
@@ -99,8 +99,12 @@ public final class WebBrowser {
      *
      * 100% 显式走 setZoomLevel(0) 而不是跳过: 浏览器实例是复用的, 玩家把缩放调回 100 时必须真的把上一次的
      * 缩放清掉。
+     *
+     * 收 double 而不是 int: 调用方传进来的已经不是配置里那个整数, 而是"配置值 x 离屏降采样比例"
+     * (见 {@link WebUiScreen#layout})。降采样必须由 zoom 等比补偿, 否则缩小离屏表面等于同时缩小了页面的
+     * CSS 视口 —— 那会挪动响应式断点、改变排版, 而本意只是降低采样密度。
      */
-    public void setZoomPercent(int percent) {
+    public void setZoomPercent(double percent) {
         MCEFBrowser b = browser;
         if (b == null) {
             return;
